ops: fold external links into the pages nav pipeline
Address review: instead of a parallel ops.externalLinks config + dedicated
useExternalNavItems/isSafeHref/render block, treat an external href as a third
destination type on a pages item, alongside layout (in-app :to) and children
(expand). OpsNavListItem now renders an item as an external anchor when it has
a safe href, an in-app router link when it has a layout, and expands when it
only has children - with the primary destination owning the row click and the
chevron demoting to an appended button.

This lets deployment links sit between dashboard pages rather than pinned below
the fixed menu items, and removes the parallel externalLinks apparatus from
nav.js and OpsNav.vue.

Layout-less leaves remain navigable (:to) as before, so no existing config
behaviour changes. Href scheme allowlist (http/https/root-relative only) is
retained to guard against dangerous schemes, and rel=noopener noreferrer is set
on external links. Adds href/target to the dashboardPage schema def.
