+++ ui/ops/src/routes/ops/nav.js
@@ -12,14 +12,6 @@ import {computed} from 'vue';
  */
 
 /**
- * @typedef {Object} ExternalNavItem
- * @property {string} title - The title of the menu item
- * @property {string} icon - The icon to display for the menu item
- * @property {string} href - The external URL the item links to
- * @property {'_self'|'_blank'} [target] - Where to open the link, defaults to '_blank'
- */
-
-/**
  * Menu Items
  * This is the main list of items
  *
@@ -109,44 +101,4 @@ export function isRouteEnabled(item) {
  */
 export function useEnabledNavItems() {
   return computed(() => navItems.filter(item => isRouteEnabled(item)));
-}
-
-/**
- * Checks whether an href is safe to render as an external link.
- *
- * Only http(s) URLs and root-relative paths are permitted. This rejects dangerous schemes
- * such as javascript: and data:, which would otherwise execute in the Ops UI origin when
- * clicked (the href comes from config, so this is defence-in-depth against a bad config).
- *
- * @param {string} href
- * @return {boolean}
- */
-function isSafeHref(href) {
-  if (typeof href !== 'string' || !href) return false;
-  // Root-relative paths (e.g. /-/ooh-ac-request/) are same-origin and safe.
-  if (href.startsWith('/')) return true;
-  try {
-    const url = new URL(href, window.location.origin);
-    return url.protocol === 'http:' || url.protocol === 'https:';
-  } catch {
-    return false;
-  }
-}
-
-/**
- * Get a computed reference to the external navigation links defined in the UI config.
- *
- * These are deployment-specific links to other applications served alongside the Ops UI
- * (e.g. standalone SPAs on the same origin). They are not router paths, so they are gated
- * purely by their presence in config rather than by pathEnabled().
- *
- * @return {ComputedRef<ExternalNavItem[]>}
- */
-export function useExternalNavItems() {
-  const uiConfig = useUiConfigStore();
-  return computed(() => {
-    const items = uiConfig.getOrDefault('ops.externalLinks', []);
-    if (!Array.isArray(items)) return [];
-    return items.filter(item => item && isSafeHref(item.href));
-  });
 }
